refactor: make changes suggested in PR, clean up code, remove comments
