+++ src/App.js
@@ -3,8 +3,6 @@ import React from 'react'
 import Game from './components/Game'
 
 function App() {
-
- 
   return (
     <div className="App">
       <header className="App-header">
