+++ src/components/Board.test.js
@@ -1,26 +1,31 @@
-import { render , screen} from "@testing-library/react";
+import { render , screen, fireEvent} from "@testing-library/react";
 import React from 'react';
 import Board from './Board';
 import GameMode from './GameMode';
 
-// beforeAll(()=>{
-//   render(<GameMode/>);
-// })
+beforeEach(()=>{
+  render(<Board
+     gridData={[1,2,3,4,5,6,7,8,9]} 
+     currentPlayerMarker='X'
+     addPlayerMarker={mockAddPlayerMarker}
+  />);
+})
+
+const mockAddPlayerMarker = (gridData, currentPlayerMarker, playerMove)=>{
+  return gridData, currentPlayerMarker, playerMove
+}
 
 describe('Board', ()=>{
   
   it('should render nine buttons', ()=>{
-    render(<Board gridData={[1,2,3,4,5,6,7,8,9]} currentPlayerMarker='X'/>);
     const gridButtonElements = screen.getAllByRole('button');
     expect(gridButtonElements.length).toBe(9);
   })
 
   it('should render a number,one in each square', async()=>{
-    render(<Board gridData={[1,2,3,4,5,6,7,8,9]} currentPlayerMarker='X'/>);
     const gridButtonElements = screen.getAllByRole('button');
     expect(gridButtonElements[0].textContent).toBe('1')
     expect(gridButtonElements[1].textContent).toBe('2')
     expect(gridButtonElements[2].textContent).toBe('3')
   })
-
 })
